+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 #### :nail_care: Polish
 
 - Allow skipping the leading pipe in variant definition with a leading constructor with an attribute. https://github.com/rescript-lang/rescript/pull/7782
+- Better error message (and recovery) when using a keyword as a record field name. https://github.com/rescript-lang/rescript/pull/7784
 
 #### :house: Internal